requirements.txt
beautifulsoup4
psycopg[binary]
requests
pytz
python-dateutil
